@@ -36,12 +36,12 @@
   },
   "devDependencies": {
     "@analogjs/vite-plugin-angular": "^1.0.0",
-    "@angular/common": "^17.0.0",
-    "@angular/compiler": "^17.0.0",
-    "@angular/compiler-cli": "^17.0.0",
-    "@angular/core": "^17.0.0",
-    "@angular/platform-browser": "^17.0.0",
-    "@angular/platform-browser-dynamic": "^17.0.0",
+    "@angular/common": "^17.3.0",
+    "@angular/compiler": "^17.3.0",
+    "@angular/compiler-cli": "^17.3.0",
+    "@angular/core": "^17.3.0",
+    "@angular/platform-browser": "^17.3.0",
+    "@angular/platform-browser-dynamic": "^17.3.0",
     "@testing-library/angular": "^15.2.0",
     "ng-packagr": "^17.3.0",
     "typescript": "5.4.2",
@@ -53,8 +53,8 @@
     "tslib": "^2.3.0"
   },
   "peerDependencies": {
-    "@angular/common": ">=16.0.0",
-    "@angular/core": ">=16.0.0"
+    "@angular/core": ">=16 < 18",
+    "@angular/common": ">=16 < 18"
   },
   "module": "build/fesm2022/tanstack-angular-form.mjs",
   "types": "build/index.d.ts",
