@@ -1,3 +1,9 @@
-export type { FieldApi, FormOptions, FieldState } from '@tanstack/form-core'
+export type {
+  FieldApi,
+  FormOptions,
+  FieldState,
+  FieldValidateFn,
+  FieldValidateAsyncFn,
+} from '@tanstack/form-core'
 export { injectForm } from './inject-form'
 export { TanStackField } from './tanstack-field.directive'
